Reorganization towards PhD thesis
Reorganized the repository to be the sister of https://github.com/CRECKMODELING/Kinetic-Mechanisms
but report the kinetic models at the status of submission of my PhD thesis.
For further developments refer to the other repository.

In general this reorganization involves:
- updated PE models
- development of PP reduced and skeletal mechanisms
- development of PS reduced and skeletal mechanisms
- introduction of validation plots in the readme files.

Probably there are issues here and there and future updates will address bug-fixes of the repository links etc