PS comprehensive update
Updated the semi-detailed mechanism to improve allyl-types description.
The resulting model has slightly worse predictions in degradation temperatures (TG profiles a bit too smooth), but has significant improvements to the volatiles yields (R^2 of 0.88 for styrene and 0.91 for dimer).
The reduced and skeletal models have been updated to reflect the improvements.
Updated also the initial species of skeletal model to be uniform with the other two.
Validated and improve thermochemistry as well.